backend/requirements.txt
fastapi
uvicorn
gunicorn
pymongo
motor
pydantic
pydantic-settings
python-dotenv
groq
langgraph
langchain
langchain-groq
langchain-community
python-multipart
websockets
